+++ imports/ui/components/Account.jsx
@@ -3,7 +3,6 @@ import { Meteor } from 'meteor/meteor';
 import { Link } from 'react-router-dom';
 import { Validators } from '/imports/api/validators/validators.js';
 
-const AddressLength = 40;
 
 export default class Account extends Component{
     constructor(props){
@@ -28,8 +27,8 @@ export default class Account extends Component{
             {fields: this.getFields() });
         if (validator)
             this.setState({
-                address: `/validator/${validator.address}`,
-                moniker: validator.description?validator.description.moniker:validator.operator_address,
+                address: `/validator/${validator?.operator_address}`,
+                moniker: validator?.description?.moniker ?? validator?.operator_address,
                 validator: validator
             });
         else
@@ -46,30 +45,14 @@ export default class Account extends Component{
             if (result){
                 // console.log(result);
                 this.setState({
-                    address: `/validator/${result.address}`,
-                    moniker: result.description?result.description.moniker:result.operator_address,
+                    address: `/validator/${result?.operator_address}`,
+                    moniker: result?.description?.moniker ?? result?.operator_address,
                     validator: result
                 });
             }
         })
     }
 
-    getAccount = () => {
-        let address = this.props.address;
-        let validator = Validators.findOne(
-            {$or: [{operator_address:address}, {delegator_address:address}, {address:address}]},
-            {fields: {address:1, description:1, operator_address:1, delegator_address:1}});
-        if (validator)
-            this.setState({
-                address: `/validator/${validator.address}`,
-                moniker: validator.description.moniker
-            });
-        else
-            this.setState({
-                address: `/validator/${address}`,
-                moniker: address
-            });
-    }
 
     componentDidMount(){
         if (this.props.sync)